Made pipeline class interface generic to any media framework
Abstract the getters for the fourCC video data format and the start of image
data in a video buffer to be non-Gstreamer specific.
Also moved the code reponsible for loading a new frame of video into a texture
which calls to the vid data start getter into a seperate virtual member
function in the GLWidget.