Update libassimp build setup in Qt project files
Remove libassimp include paths from project files, now uses pkg-config
to get compile and linker flags on the host build and in the target
sysroot when cross compiling using the PKG_CONFIG_PATH set in the
environment.
So ensure that libassimp is correctly installed on host for desktop
build or target sysroot for cross compiled targets.