TI OMAP3EVM streaming textures with bc-cat mostly working
Video can be seen as a texture on some of the model now, with GL_NEAREST attributes.
The shader for texture on the quad is not working yet, just rendered as a solid colour.

Initial code to use bc-cat driver was wrong, add a pool of textures for each video,
when a new buffer for a video is received, just need to set the active texture ID to
the ID from the pool which correlates to the buffer which has a new video frame in it.
That texture ID is then bound when drawing as normal in the base GLWidget class.
Add map containter instance to handle buffer index to texture ID matching.

Control how many buffers are active per video dynamically, when a buffer address is
received a second time, assume number of different buffers received so far is the total
for that video.

Fill bc_buf_params_t width+height based on number of pixels, not byte size.

Add extra debug logging to help get this far.

Add abstracted fourCC values, macro to convert to BC equivalents (Just a NOP).